Move to math/utils.py
